backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
qdrant-client>=1.7.0
numpy>=1.21.0
psycopg2-binary
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
better-exceptions==0.3.2
cryptography==41.0.8
httpx==0.25.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-slugify==8.0.1
aiofiles==23.2.1
beautifulsoup4==4.12.2
requests==2.31.0
markdown==3.5.1
python-frontmatter==1.1.0